# Fix microscopy ground-truth range of values