## Checklist

#### GitHub

- [x] I've given this PR a concise, self-descriptive, and meaningful title
- [x] I've linked relevant issues in the PR body
- [x] I've applied [the relevant labels](https://www.neuro.polymtl.ca/software/contributing#pr_labels) to this PR
- [x] I've assigned a reviewer

#### PR contents

- [x] I've consulted [ivadomed's internal developer documentation](https://github.com/ivadomed/ivadomed/wiki) to ensure my contribution is in line with any relevant design decisions
- [x] I've added [relevant tests](https://github.com/ivadomed/ivadomed/wiki/tests) for my contribution
- [x] I've updated the [relevant documentation](https://github.com/ivadomed/ivadomed/wiki/documentation) for my changes, including argparse descriptions, docstrings, and ReadTheDocs tutorial pages

## Description
This PR fix an issue where the Dice score would be >1 when training with microscopy data. In the loader, the PNG/TIF ground-truth values of 0 and 255 are now converted to 0 and 1 before converting to Nibabel/NifTI (as it is assumed for MRI data). These Nibabel/NifTI are then used for training, testing and evaluation and the values will have the same 0-1 range everywhere.

Additionally, the PNG/TIF values are now read and written as uint8 in the NifTI file to save space, both for the images and ground-truths.

## Questions
- In #868 and #468, we talked about checking the range of the input values for the metrics. Wouldn't it apply to the other metrics as well? If so, where should we do the verification? Feedback welcome!
- I also ran into an issue while trying to check the range in `dice_score`. The prediction files (`_pred.nii.gz`) sometimes have values slightly over 1 (!), ex: 1.0000000591389835. It happens both with microscopy and MRI. It may be a precision thing but we can't directly check for the [0-1] range.

## Linked issues
Fixes #868