@hermancollin 2021-07-22
While testing this PR I still encountered dice scores superior to 1 (1.30 for class0 and 1.85 for class1). Adding a debut print inside the `ivadomed.metrics.py:dice_score()` function still gives me this when running a test:
```
dice_score() - im1.max(): 255.0
dice_score() - im2.max(): 1.0
dice_score() - im1.max(): 255.0
dice_score() - im2.max(): 1.0
```
To reproduce the issue with dice scores >1, I put an example on `duke/temp/arcol/to_test_pr_870`.