@hermancollin 2021-07-22
Ok I found out why it wasn't working on my side. I hadn't deleted the nii.gz files so although the conversion to the [0,1] range was working, the nifti files weren't saved by the `ivadomed.loader.segmentation_pair.py:convert_file_to_nifti()` function. Now, I get the expected 0.7002 and 0.7971 dice scores for classes 0 and 1. Great work :tada: 

I did however encounter the problem you mentionned with max values in predictions slightly over 1 (1.0000000591389835... exactly the same value @mariehbourget reported)