@mariehbourget 2021-07-22
Thanks @hermancollin for the debugging!

> I did however encounter the problem you mentionned with max values in predictions slightly over 1 (1.0000000591389835... exactly the same value @mariehbourget reported). This happened in evaluation but not in inference.

The evaluation step reads the NifTI file written previously at the testing step. I checked the `_pred.nii.gz` file and the value is indeed 1.0000000591389835, even though it prints as 1.0 just before saving the NifTI here 
https://github.com/ivadomed/ivadomed/blob/7a595ae39e94085837582f7176b8ec59bd9a2bc9/ivadomed/inference.py#L198

I ran some more tests:

1.  With single-class model on microscopy or MRI `pred_masks`:
- The max value is 1 (int) when using `binarize_prediction` with threshold 0.5
- The max value is 1.0000000591389835 without binarization

2. With multi-class model on microscopy or MRI `pred_masks`:
- The max value is 1.0000000591389835 with or without `binarize_maxpooling`

So it seems that `binarize_maxpooling` doesn't return the value as an integer and when values are in float, 1.0 becomes 1.0000000591389835 in the NifTI file. Interesting...