@hermancollin 2021-07-22
> 1. With multi-class model on microscopy or MRI `pred_masks`:
> 
> * The max value is 1.0000000591389835 with or without `binarize_maxpooling`
> 
> So it seems that `binarize_maxpooling` doesn't return the value as an integer and when values are in float, 1.0 becomes 1.0000000591389835 in the NifTI file. Interesting...

@mariehbourget correct me if I'm wrong but shouldn't the `binarize_maxpooling` postprocess return integers instead of floats?