@mariehbourget 2021-07-22
> @mariehbourget correct me if I'm wrong but shouldn't the `binarize_maxpooling` postprocess return integers instead of floats?

I think so to, right now it is returned as a float that we could convert as an int [here](https://github.com/ivadomed/ivadomed/blob/7a595ae39e94085837582f7176b8ec59bd9a2bc9/ivadomed/postprocessing.py#L300) but it would be good to know if there was a reason for the float in the first place @andreanne-lemay, @charleygros?

> Although this would fix the current problem, the issue would persist when no postprocessing is done on the predictions...

Yes, this would require more investigation, the value is 1.0 before the NifTI file is saved. Maybe there is a change in data type that affects the precision?