@mariehbourget 2021-07-28
As discussed in meeting, I may have found an issue with this PR where I have "not as good" result with training on this branch as I had on July 5th on master (dice went down from ~0.85 to 0.75 with similar config). I'm investigating now if it's due to the 8 bit conversion or the GT values of 0 and 1, and I'll keep you posted when ready for review.

In the meantime, I would like to have some feedback on the value of `1.0000000591389835` found in prediction files that are not binarized or binarized with maxpooling, see this [comment](https://github.com/ivadomed/ivadomed/pull/870#issuecomment-885155963).