@charleygros 2021-07-28
> I think so to, right now it is returned as a float that we could convert as an int [here](https://github.com/ivadomed/ivadomed/blob/7a595ae39e94085837582f7176b8ec59bd9a2bc9/ivadomed/postprocessing.py#L300) but it would be good to know if there was a reason for the float in the first place @andreanne-lemay, @charleygros?

I do think it would be ok to have a int-conversion here.