@hermancollin 2021-07-29
> As discussed in meeting, I may have found an issue with this PR where I have "not as good" result with training on this branch as I had on July 5th on master (dice went down from ~0.85 to 0.75 with similar config). I'm investigating now if it's due to the 8 bit conversion or the GT values of 0 and 1, and I'll keep you posted when ready for review.

I may have found one cause of the problem here. Using @mariehbourget's config file on the SEM microscopy dataset, the following problem occurs:  the SEM dataset contains some values other than 0 or 255 in the ground-truth (most likely a compression artifact as they were intended to be "hard" segs). The `img = img // 255` may be too simple to fix the problem because values such as 130, 218 and 233 in the dataset turn to 0s with this floor division. 

In general, I don't think this is a behavior we want for the loader. However, looking at the color profiles of the troublesome masks, the % of pixel anomaly is ridiculously low so this definitely doesn't explain the training performance difference @mariehbourget reported.