@mariehbourget 2021-08-05
>  The `img = img // 255` may be inappropriate to fix the problem because values such as 130, 218 and 233 in the dataset turn to 0s with this floor division.
In general, I don't think this is a behavior we want for the loader. However, looking at the color profiles of the troublesome masks, the % of pixel anomaly is ridiculously low so this definitely doesn't explain the training performance difference

Thank you @hermancollin for checking this. I agree with you and changed the floor division by a thresholding at 0.5. In theory, ground-truth should already be 0 and 255 with no intermediate values. But if that is not the case, a threshold makes more sense than a floor division.

> Hypothesis: There seems to be conversion loss somewhere. On my side, everytime a prediction is saved, I have a lossy conversion warning (float64 to uint8). When forcing GT values to 0 or 1 this early in the training pipeline, maybe we suffer from more conversion loss in the validation/testing step.
This conversion from float64 to uint8 of each prediction is normal, it happens when the final Nibabel predictions (in float64) is written to PNG (uint8).

Also, I have found the source of the discrepancies between training on master and training on this branch and will document it here in a separate comment as soon as possible.