@mariehbourget 2021-08-05
This PR is ready for review.
There was a few questions along the way with discrepancies between training on master and training on this branch. I have now found the sources for these discrepancies and will document them here.
It's a long story, please bear with me.

### What the PR does:
- The PNG/TIF values (8 bit between 0 and 255) that were previously stored as `float32`, are now read and written as `uint8` in the NifTI files to save space, both for the images and ground-truths.
- The PNG/TIF ground-truth values of range 0-255 are now converted to 0s and 1s to fix an issue where the Dice score would be >1 when training with microscopy data. This is exactly the same as MRI data using 0s and 1s in `uint8` for their ground-truth. Additionally, the range 0-255 for the ground-truth was incompatible with the softseg pipeline.
- I added a test in test_loader.py to check the conversion of 0s and 1s values.

### First issue with training:
- I reported previously "not as good" result with training on this branch as I had on master (Dice went down from ~0.85 to 0.75 with similar config).
- After investigation, this was due to a poor management of the split dataset on my part. The split was vastly different between my two batches of tests causing the big differences in Dice.

### Second issue with training:
- When using the exact same split and config, I still encountered some minor systematic differences in Dice. I ran each test 4 times with always the same results.
    - On master (2cdc40ad and 0a6fffd): `Training Dice ~0.86` and `Validation Dice ~0.85`
    - On this branch (4523a0f): `Training Dice ~ 0.85` and `Validation Dice ~0.83-0.84`
- I thought it could be due to bad conversion between `uint8` and `float` data somewhere in the pipeline, but it's not the case. Results  on this branch are similar with either formats and I checked the values throughout the code. I can confirm that my issue had nothing to do with the datatype and that no bad conversion happens with either `imageio` (to read PNG) or `Nibabel` (to read/write Nifti).
- So the differences had to be because of the 0-1 range for ground-truths. I finally found the difference in the thresholding for the ground-truth that happens after the pre-processing and data augmentation [here](https://github.com/ivadomed/ivadomed/blob/4460d1379b77ffd114d4b416fda0924b40a51e3e/ivadomed/loader/mri2d_segmentation_dataset.py#L216).
- At this point, all data (regardless of how it is stored in the NIfTI file) is in `float32` in ivadomed.
- But the threshold for binarizing the intermediate values of ground-truth after the data augmentation is hardcoded to 0.5 (which makes sense when the ground-truths are initially 0s and 1s).
- So the differences between training on master and training on this branch came from the threshold being applied at a different "place" in the values range i.e. `0.5/255` vs. `0.5/1`.
- I confirmed this by running trainings with a threshold of `0.00196/1` (equivalent of `0.5/255`), and I obtain the same results as I had on master.

### Conclusion:
- Converting ground-truth to 0s and 1s when reading the PNG files and converting to NIfti files is essential to ensure the rest of the pipeline is applied correctly and metrics computed properly.
- Writing the NIfTI files in uint8 of float32 has no incidence on the rest of the pipeline.
- Differences were due to the threshold after data augmentation of 0.5 being wrongly applied on 0-255 values on master instead of  0-1 values on this branch.
- Even though the training/validation Dice are slight lower with 0-1 values, testing Dice on my test image are similar if not better with with 0-1.
- We could discuss about adding a parameter to the config and optimizing the threshold for binarization that is currently hardcoded at 0.5, but this is out of scope for this PR.

### About the `1.0000000591389835` value in NIftI predictions:
- As discussed previously in this PR [here](https://github.com/ivadomed/ivadomed/pull/870#issuecomment-885155963), we found a value >1 on some NIfTI predictions files.
- This seems to be a precision error when the files are written.
- I also decided that this subject is out of scope for this PR and will open a separate issue on the matter.